+++ src/main/java/kuit/project/beering/controller/MemberController.java
@@ -6,9 +6,9 @@ import kuit.project.beering.dto.request.member.MemberLoginRequest;
 import kuit.project.beering.dto.request.member.MemberSignupRequest;
 import kuit.project.beering.dto.response.member.MemberLoginResponse;
 import kuit.project.beering.service.MemberService;
-import kuit.project.beering.util.AgreementValidationException;
+import kuit.project.beering.util.exception.AgreementValidationException;
 import kuit.project.beering.util.BaseResponse;
-import kuit.project.beering.util.FieldValidationException;
+import kuit.project.beering.util.exception.FieldValidationException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.BindingResult;
