+++ src/main/java/kuit/project/beering/service/MemberService.java
@@ -12,7 +12,7 @@ import kuit.project.beering.repository.MemberRepository;
 import kuit.project.beering.security.auth.AuthMember;
 import kuit.project.beering.security.jwt.JwtInfo;
 import kuit.project.beering.security.jwt.JwtTokenProvider;
-import kuit.project.beering.util.DuplicateUsernameException;
+import kuit.project.beering.util.exception.DuplicateUsernameException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
