--- src/main/java/kuit/project/beering/util/AgreementValidationException.java
@@ -1,11 +1,0 @@
-package kuit.project.beering.util;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.validation.BindingResult;
-
-@AllArgsConstructor
-@Getter
-public class AgreementValidationException extends RuntimeException {
-    private BindingResult bindingResult;
-}
