+++ src/main/java/kuit/project/beering/util/exception/AgreementValidationException.java
@@ -1,16 +1,11 @@
-package kuit.project.beering.util;
+package kuit.project.beering.util.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.validation.BindingResult;
 
-/**
- * @Brief 회원 가입 검증 예외
- */
-@Getter
 @AllArgsConstructor
-public class FieldValidationException extends RuntimeException {
-
+@Getter
+public class AgreementValidationException extends RuntimeException {
     private BindingResult bindingResult;
-
 }
