+++ src/main/java/kuit/project/beering/util/exception/CustomJwtException.java
@@ -1,6 +1,7 @@
-package kuit.project.beering.util;
+package kuit.project.beering.util.exception;
 
 import io.jsonwebtoken.JwtException;
+import kuit.project.beering.util.BaseResponseStatus;
 import lombok.Getter;
 
 @Getter
