+++ src/main/java/kuit/project/beering/util/exception/DuplicateUsernameException.java
@@ -1,4 +1,4 @@
-package kuit.project.beering.util;
+package kuit.project.beering.util.exception;
 
 public class DuplicateUsernameException extends RuntimeException {
 }
